# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.LegacyStatsResult:
    properties:
      abbreviations:
        type: integer
      cached_at:
        type: string
      calculators:
        type: integer
      consultants:
        type: integer
      drugs:
        type: integer
      faqs:
        type: integer
      health_facilities:
        type: integer
      medical_guidelines:
        type: integer
      ministry_directory:
        type: integer
      success:
        type: boolean
      total_users:
        type: integer
      unread_messages_count:
        type: integer
      user_conversations_count:
        type: integer
    type: object
  handlers.LegacyTreeResult:
    properties:
      data:
        items:
          $ref: '#/definitions/services.TreeNode'
        type: array
      level:
        type: integer
      success:
        type: boolean
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: admin@mediguide.local
        type: string
      password:
        example: Admin123!
        type: string
    type: object
...
  handlers.LogoutResult:
    properties:
      logged_out:
        example: true
        type: boolean
    type: object
...
info:
  contact:
    email: admin@mediguide.local
    name: MediGuide Backend
  description: Offline-first clinical guideline backend API for MediGuide.
  license:
    name: Proprietary
  termsOfService: https://mediguide.local/terms
  title: MediGuide Backend API
  version: "1.0"
paths:
  /api/v1/consultants/tree:
    get:
      description: Legacy v1 endpoint that groups consultants by region, city, then
        specialty.
      parameters:
      - description: Tree level
        in: query
        maximum: 2
        minimum: 0
        name: level
        type: integer
      - description: JSON encoded filters
        in: query
        name: filters
        type: string
      - description: Optional JSON context
        in: query
        name: context
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LegacyTreeResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get consultants tree
      tags:
      - legacy-v1
...
  /api/v1/ministry-directory/tree:
    get:
      description: Legacy v1 endpoint that groups ministry directory contacts by region,
        district, then ministry.
      parameters:
      - description: Tree level
        in: query
        maximum: 2
        minimum: 0
        name: level
        type: integer
      - description: JSON encoded filters
        in: query
        name: filters
        type: string
      - description: Optional JSON context
        in: query
        name: context
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LegacyTreeResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get ministry directory tree
      tags:
      - legacy-v1
  /api/v1/overview:
    get:
      description: Dashboard-oriented legacy v1 overview endpoint with metrics, pipeline,
        engagement, support and coverage totals.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LegacyOverviewResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get legacy overview metrics
      tags:
      - legacy-v1
  /api/v1/stats:
    get:
      description: Legacy v1 stats endpoint used by the mobile home screen. Authentication
        is optional, but user-specific message counts are only returned when a valid
        bearer token is provided.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LegacyStatsResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get legacy mobile stats
      tags:
      - legacy-v1
  /api/v2/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT bearer token.
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Log in a user
      tags:
      - auth
  /api/v2/auth/logout:
    post:
      description: Revoke the current authenticated session.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out current session
      tags:
      - auth
  /api/v2/auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset confirmation
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Confirm a password reset
      tags:
      - auth
  /api/v2/auth/password-reset/request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AccountActionResult'
      summary: Request a password reset
      tags:
      - auth
  /api/v2/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token and rotated refresh
        token.
      parameters:
      - description: Refresh payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Refresh a user session
      tags:
      - auth
  /api/v2/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new backend user account.
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a user
      tags:
      - auth
  /api/v2/calculator-usage/{usageId}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Usage session ID
        format: uuid
        in: path
        name: usageId
        required: true
        type: string
      - description: Usage completion
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/services.FinishCalculatorUsageInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CalculatorUsageEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finish a calculator usage session
      tags:
      - calculators
...
  /api/v2/calculators/{id}:
    delete:
      parameters:
      - description: Calculator ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Archive a calculator or decision tool
      tags:
      - calculators
    get:
      parameters:
      - description: Calculator ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CalculatorEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a calculator or decision tool
      tags:
      - calculators
    patch:
      consumes:
      - application/json
      parameters:
      - description: Calculator ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Calculator fields
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCalculatorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CalculatorEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a calculator or decision tool
      tags:
      - calculators
...
  /api/v2/drugs/{id}/usage:
    post:
      parameters:
      - description: Drug ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.DrugUsageEnvelope'
      security:
      - BearerAuth: []
      summary: Record authenticated drug usage
      tags:
      - drugs
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer access token. Example: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"